Add roles authorization for logs

+++ RadiometerWebApp/Controllers/AccountController.cs
@@ -68,7 +68,7 @@ public class AccountController : Controller
             var claims = new List<Claim>
             {
                 new Claim(ClaimsIdentity.DefaultNameClaimType, user.Login),
-                new Claim(ClaimsIdentity.DefaultRoleClaimType, ((int)user.Role).ToString())
+                new Claim(ClaimsIdentity.DefaultRoleClaimType, user.Role.ToString())
             };
             ClaimsIdentity? claimsIdentity =
                 new ClaimsIdentity(claims, "Token", ClaimsIdentity.DefaultNameClaimType,
